♻️ 処理をファイル分割
